avoid circular dependencies
