+++ shared/src/formEstablishment/FormEstablishmentDtoBuilder.ts
@@ -147,6 +147,7 @@ export const defaultValidFormEstablishment: FormEstablishmentDto = {
   },
   acquisitionKeyword: undefined,
   acquisitionCampaign: undefined,
+  isBanned: false,
 };
 
 export const fullyUpdatedFormEstablishment: FormEstablishmentDto = {
@@ -200,6 +201,7 @@ export const fullyUpdatedFormEstablishment: FormEstablishmentDto = {
   fitForDisabledWorkers: "no",
   isEngagedEnterprise: false,
   nextAvailabilityDate: new Date("2025-02-01").toISOString(),
+  isBanned: false,
 };
 
 const emptyFormEstablishment: FormEstablishmentDto = {
@@ -227,6 +229,7 @@ const emptyFormEstablishment: FormEstablishmentDto = {
   },
   acquisitionKeyword: undefined,
   acquisitionCampaign: undefined,
+  isBanned: false,
 };
 
 export class FormEstablishmentDtoBuilder
